The connector's tool calls were never written down, and I called the rest „every call"
The tester's observation: mcpServer.runTool writes the cost ledger and not
tool_call_log. It could not — thread_id was NOT NULL and a connector call has
no conversation — so the one table that records what a tool DID held the
chat's half of the product and nothing else. For three days I took counts out
of it and said „every call" when I had „every chat call". Third time in two
days that a table showed me a part and I read it as the whole.

  - migration 166: thread_id nullable, surface added, default 'chat'. Every
    existing row is a chat call by construction, so no backfill
  - surface is stored, not derived from a null thread: a meaning that has to be
    remembered gets forgotten, and the counts either side of today are only
    comparable through it
  - runTool records both ways out — a log of the calls that worked answers „how
    often does this fail" with silence
  - the arguments really are passed: a fourth parameter, at all 62 call sites.
    A blank args_summary cannot be told apart from „nothing was passed"
  - same redaction as the chat's: phone as last four digits (D149), free text
    counted and never copied
  - the test reads the source, like registryParity next door, and holds the
    parse to the file's own count rather than a floor — a unit test cannot see
    a caller that never calls, which is exactly how this survived

Also: participant_count and asked_count on /admin/chorus/campaigns were going
out as strings (Postgres COUNT is bigint), so "12" > "9" was false in the
client. I warned the frontend and left the server sending the wrong type,
which is the wrong half to fix. ::int.

+++ src/api/routes/admin.routes.ts
@@ -3641,8 +3641,17 @@ adminRouter.get('/chorus/campaigns', async (req: Request, res: Response) => {
       query(
         `SELECT c.id, c.target_phone, c.target_label, c.city, c.status, c.ask_count_dial,
                 c.opened_at, c.closed_at, c.closed_reason,
-                COUNT(p.id) AS participant_count,
-                COUNT(p.id) FILTER (WHERE p.asked_at IS NOT NULL) AS asked_count,
+                -- ::int, and it is not cosmetic. Postgres COUNT is bigint, and
+                -- node-postgres hands a bigint back as a STRING rather than
+                -- lose precision on it. So this page was sending "12" and "9",
+                -- and in the client "12" > "9" is false — a sort or a
+                -- threshold on either column reads backwards, silently, on
+                -- exactly the rows that matter most. I warned the frontend
+                -- about it and left the server sending the wrong type, which
+                -- is the wrong half to fix: a count of campaign participants
+                -- has no precision to lose.
+                COUNT(p.id)::int AS participant_count,
+                COUNT(p.id) FILTER (WHERE p.asked_at IS NOT NULL)::int AS asked_count,
                 MIN(p.scheduled_ask_at) FILTER (WHERE p.asked_at IS NULL)
                   AS next_ask_due_at,
                 COALESCE(
